chore: add measurement transform test and align type
In a follow up we need to align al the respective logic for measurements to remove any ambiguity when it comes to how many measurement/measurement names can be added to one object.

It's only one!

The current structure which leaves that open was never used and is a historical thingy where the first goals was to be as open as possible to different use cases.