Clean up batch processing
After some suggestions by @bfmatei, there might be cases of transports
that require instantaneous execution of the signal (such as a WebSocket,
a Console or any stream). To that end, this commit removes the
deprecation notice of the un-batched transports and instead changes the
core logic which is responsible for sending the item.